pyproject.toml
[project]
name = "uv-whisper-py13"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.44.1",
    "openai-whisper>=20250625",
    "python-dotenv>=1.1.1",
]
